ai-web/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai>=0.8.0
faiss-cpu
numpy
